MED-80 chore: refactor redux toolkit
