--- apps/medon-fe/src/redux/features/backend/types.ts
@@ -1,8 +1,0 @@
-export interface IResetPassword {
-  token?: string;
-  newPassword: string;
-}
-
-export interface IForgetPassword {
-  email: string;
-}
